res/requirements.txt
PyYaml
argparse